move residuals into old gmm_utils
